pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mini_muse"
version = "0.1.0"
description = "AI-powered image generation and feedback application"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "mini_muse", email = "mini_muse@example.com"}
]
keywords = ["ai", "image-generation", "stable-diffusion", "ollama", "tui"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # TUI Framework
    "textual>=0.41.0",
    "rich>=13.0.0",
    
    # Image Generation
    "diffusers>=0.21.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.30.0",
    "accelerate>=0.20.0",
    
    # Image Processing
    "pillow>=10.0.0",
    "opencv-python>=4.8.0",
    
    # LLM Integration
    "ollama>=0.1.7",
    
    # Configuration and Data
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "toml>=0.10.2",
    
    # Async and Utilities
    "aiofiles>=23.0.0",
    "httpx>=0.24.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
mini-muse = "mini_muse.main:main"

[project.urls]
Homepage = "https://github.com/mini_muse/mini_muse"
Repository = "https://github.com/mini_muse/mini_muse"
Issues = "https://github.com/mini_muse/mini_muse/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["mini_muse*"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["mini_muse"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"